First working code for Elastic
Added a basic code to dump Packets from Pcap file into Elastic. The idea
was just to make sure we are using basic APIs correctly etc. Also added
a `dyn_mappings.json` file in the same directory. This is just a
starting point, eventually the plan is to make these mappings getting
better so that we can have them better optimized!